Added get_week and change_username functions
On the tutor side, added get_week() to dynamically determine the week number as per NUS Academic Calendar. This is necessary for the Column Name. 

On the student side, added change_username(matric_no) to update a student's changed telegram username in our database. 

Also removed unnecessary imports and added a few TODOs in the comments.